Support Python 2.7
Originally the application was only intended to be used with Python 3
and made use of statements and syntaxes that are not valid in Python 2.
It seems however, that the days of Python 2 are not over and some users
had interest in using the application with Python 2.

With the same codebase, the application now runs using Python 2.7, 3.3
and 3.4. This should close #1.

Signed-off-by: Cilyan Olowen <gaknar@gmail.com>